Add Yubikey pairing functionality to bootstrap-mac role
Added a new set of tasks for handling Yubikey pairing in the bootstrap-mac role. A new flag (`bind_yk`) has been introduced in the defaults to control the execution of these tasks. When `bind_yk` is true, the ansible playbook will ensure the Yubikey is correctly paired with the macOS.