Added caffeinate command to tunnels.sh script
The 'caffeinate -d' command has been added to the tunnels.sh script in the bootstrap-mac role. This command will prevent the system from sleeping while the script is running, ensuring uninterrupted execution of the script when ssh tunnels are established. The command will only be added if there are ssh tunnels to be established.